# src/swagger.yaml
openapi: 3.0.0
info:
  title: AI Sales API
  description: Complete API documentation for the AI Sales system.
  version: 1.0.0
security:
  - BearerAuth: []
paths:
...
  /api/v1/roles:
    post:
      tags: [Roles]
      summary: Create a new role
      description: Create a new role in the system.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreate'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: Role with this name already exists

    get:
      tags: [Roles]
      summary: Get all roles
      description: Retrieve all roles in the system.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
  /api/v1/roles/{role_id}:
    get:
      tags: [Roles]
      summary: Get role by ID
      description: Retrieve a role by its unique ID.
      security:
        - BearerAuth: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Role retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '404':
          description: Role not found

    put:
      tags: [Roles]
      summary: Update role
      description: Update role details by its unique ID.
      security:
        - BearerAuth: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdate'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '404':
          description: Role not found

    delete:
      tags: [Roles]
      summary: Delete role
      description: Delete a role by its unique ID.
      security:
        - BearerAuth: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Role deleted successfully
        '404':
          description: Role not found

  # Role Permissions Routes
...
  /api/v1/interaction-modes:
    get:
      tags: [Interaction Modes]
      summary: Get all interaction modes
      description: Retrieve all interaction modes.
      responses:
        '200':
          description: Interaction modes retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InteractionModeResponse'
    post:
      tags: [Interaction Modes]
      summary: Create a new interaction mode
      description: Create a new interaction mode.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionModeCreate'
      responses:
        '201':
          description: Interaction mode created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModeResponse'
  /api/v1/interaction-modes/{mode_id}:
    get:
      tags: [Interaction Modes]
      summary: Get interaction mode by ID
      description: Retrieve an interaction mode by its ID.
      parameters:
        - name: mode_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Interaction mode retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModeResponse'
        '404':
          description: Interaction mode not found
...
  /api/v1/ai-roles/:
    post:
      tags: [AI Roles]
      summary: Create a new AI role
      description: Create a new AI role in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRoleCreate'
      responses:
        '201':
          description: AI role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIRoleResponse'
    get:
      tags: [AI Roles]
      summary: Get all AI roles
      description: Retrieve all AI roles in the system.
      responses:
        '200':
          description: AI roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AIRoleResponse'
  /api/v1/ai-roles/{ai_role_id}:
    get:
      tags: [AI Roles]
      summary: Get AI role by ID
      description: Retrieve an AI role by its unique ID.
      parameters:
        - name: ai_role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: AI role retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIRoleResponse'
        '404':
          description: AI role not found
    put:
      tags: [AI Roles]
      summary: Update AI role
      description: Update an AI role by its unique ID.
      parameters:
        - name: ai_role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRoleUpdate'
      responses:
        '200':
          description: AI role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIRoleResponse'
        '404':
          description: AI role not found
    delete:
      tags: [AI Roles]
      summary: Delete AI role
      description: Delete an AI role by its unique ID.
      parameters:
        - name: ai_role_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: AI role deleted successfully
        '404':
          description: AI role not found

  # Industries Routes
...
  /api/v1/manufacturing-models/:
    post:
      tags: [Manufacturing Models]
      summary: Create a new manufacturing model
      description: Create a new manufacturing model in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturingModelCreate'
      responses:
        '201':
          description: Manufacturing model created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManufacturingModelResponse'
    get:
      tags: [Manufacturing Models]
      summary: Get all manufacturing models
      description: Retrieve all manufacturing models in the system.
      responses:
        '200':
          description: Manufacturing models retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManufacturingModelResponse'
  /api/v1/manufacturing-models/{manufacturing_model_id}:
    get:
      tags: [Manufacturing Models]
      summary: Get manufacturing model by ID
      description: Retrieve a manufacturing model by its unique ID.
      parameters:
        - name: manufacturing_model_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Manufacturing model retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManufacturingModelResponse'
        '404':
          description: Manufacturing model not found
    put:
      tags: [Manufacturing Models]
      summary: Update manufacturing model
      description: Update a manufacturing model by its unique ID.
      parameters:
        - name: manufacturing_model_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturingModelUpdate'
      responses:
        '200':
          description: Manufacturing model updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManufacturingModelResponse'
        '404':
          description: Manufacturing model not found
    delete:
      tags: [Manufacturing Models]
      summary: Delete manufacturing model
      description: Delete a manufacturing model by its unique ID.
      parameters:
        - name: manufacturing_model_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Manufacturing model deleted successfully
        '404':
          description: Manufacturing model not found

  # Plant Size Impacts Routes
...
  /api/v1/interaction-mode-ai-roles:
    post:
      tags: [Interaction Mode AI Roles]
      summary: Create an Interaction Mode AI Role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionModeAIRoleBase'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModeAIRoleResponse'
    get:
      tags: [Interaction Mode AI Roles]
      summary: Get all Interaction Mode AI Roles
      responses:
        '200':
          description: List of Interaction Mode AI Roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InteractionModeAIRoleResponse'
  /api/v1/interaction-mode-ai-roles/{id}:
    get:
      tags: [Interaction Mode AI Roles]
      summary: Get Interaction Mode AI Role by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Interaction Mode AI Role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModeAIRoleResponse'
        '404':
          description: Not found
    put:
      tags: [Interaction Mode AI Roles]
      summary: Update Interaction Mode AI Role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionModeAIRoleBase'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModeAIRoleResponse'
        '404':
          description: Not found
    delete:
      tags: [Interaction Mode AI Roles]
      summary: Delete Interaction Mode AI Role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /api/v1/interaction-mode-manufacturing-models:
    post:
      tags: [Interaction Mode Manufacturing Models]
      summary: Create an Interaction Mode Manufacturing Model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionModeManufacturingModelBase'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModeManufacturingModelResponse'
    get:
      tags: [Interaction Mode Manufacturing Models]
      summary: Get all Interaction Mode Manufacturing Models
      responses:
        '200':
          description: List of Interaction Mode Manufacturing Models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InteractionModeManufacturingModelResponse'
  /api/v1/interaction-mode-manufacturing-models/{id}:
    get:
      tags: [Interaction Mode Manufacturing Models]
      summary: Get Interaction Mode Manufacturing Model by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Interaction Mode Manufacturing Model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModeManufacturingModelResponse'
        '404':
          description: Not found
    put:
      tags: [Interaction Mode Manufacturing Models]
      summary: Update Interaction Mode Manufacturing Model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionModeManufacturingModelBase'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModeManufacturingModelResponse'
        '404':
          description: Not found
    delete:
      tags: [Interaction Mode Manufacturing Models]
      summary: Delete Interaction Mode Manufacturing Model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RoleBase:
      type: object
      properties:
        name:
          type: string
          enum:
            - sales_person
            - manager
            - admin
            - super_admin
        description:
          type: string
          nullable: true
    RoleCreate:
      allOf:
        - $ref: '#/components/schemas/RoleBase'
    RoleUpdate:
      type: object
      properties:
        name:
          type: string
          enum:
            - sales_person
            - manager
            - admin
            - super_admin
          nullable: true
        description:
          type: string
          nullable: true
    RoleResponse:
      allOf:
        - $ref: '#/components/schemas/RoleBase'
        - type: object
          properties:
            role_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            status_active:
              type: boolean

    # Role Permission Schemas
...
    InteractionModeCreate:
      type: object
      required:
        - name
        - description
        - prompt_template
      properties:
        name:
          type: string
          enum: [prospecting, sales, closing]
        description:
          type: string
        prompt_template:
          type: string
    InteractionModeResponse:
      type: object
      properties:
        mode_id:
          type: string
          format: uuid
        name:
          type: string
          enum: [prospecting, sales, closing]
        description:
          type: string
        prompt_template:
          type: string
        status_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AIRoleBase:
      type: object
      properties:
        name:
          type: string
    AIRoleCreate:
      allOf:
        - $ref: '#/components/schemas/AIRoleBase'
    AIRoleUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
    AIRoleResponse:
      allOf:
        - $ref: '#/components/schemas/AIRoleBase'
        - type: object
          properties:
            ai_role_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...
    ManufacturingModelBase:
      type: object
      properties:
        name:
          type: string
    ManufacturingModelCreate:
      allOf:
        - $ref: '#/components/schemas/ManufacturingModelBase'
    ManufacturingModelUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
    ManufacturingModelResponse:
      allOf:
        - $ref: '#/components/schemas/ManufacturingModelBase'
        - type: object
          properties:
            manufacturing_model_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...
    InteractionModeAIRoleBase:
      type: object
      properties:
        mode_id:
          type: string
        ai_role_id:
          type: string
        prompt_template:
          type: string
    InteractionModeAIRoleResponse:
      allOf:
        - $ref: '#/components/schemas/InteractionModeAIRoleBase'
        - type: object
          properties:
            interaction_mode_ai_role_id:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    InteractionModeManufacturingModelBase:
      type: object
      properties:
        mode_id:
          type: string
        manufacturing_model_id:
          type: string
        prompt_template:
          type: string
    InteractionModeManufacturingModelResponse:
      allOf:
        - $ref: '#/components/schemas/InteractionModeManufacturingModelBase'
        - type: object
          properties:
            interaction_mode_manufacturing_model_id:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...